Fix too early abort in pre-run when installing types while
If the mypy configuration specifies files to test, mypy
will also test files in pre-run when installing types.

With set -euxv, the error exit from such pre-run will
terminate the script early before runing reviewdog!

It's not needed, just add the flags:

With current mypy versions, the pre-run is not needed and
installing types can be done in the regular analysis run.

Signed-off-by: Bernhard Kaindl <bernhardkaindl7@gmail.com>